move `UpgradeModal` to `/common/modals`
