+++ clients/admin-ui/src/features/config-wizard/AddSystem.tsx
@@ -19,7 +19,7 @@ import {
   ADD_SYSTEMS_MANUAL_ROUTE,
   ADD_SYSTEMS_MULTIPLE_ROUTE,
 } from "~/features/common/nav/v2/routes";
-import { UpgradeModal } from "~/features/common/UpgradeModal";
+import { UpgradeModal } from "~/features/common/modals/UpgradeModal";
 import { ValidTargets } from "~/types/api";
 
 import { changeStep, setAddSystemsMethod } from "./config-wizard.slice";
